Decouple Yosys netlist build from tsfpga.vivado package
The tsfpga/yosys package previously imported BuildResult, generic value
classes, and copy_and_combine_dicts from tsfpga.vivado.*, creating an
unwanted dependency on the Vivado backend for a Yosys-only build flow.

- Move BuildResult to tsfpga/build_result.py. The only genuinely
  Vivado-specific bit (maximum_logic_level's use of
  LogicLevelDistributionParser) is now a deferred import, so it has no
  cost unless logic_level_distribution is actually used.
- Move StringGenericValue/BitVectorGenericValue to tsfpga/generics.py.
- Move copy_and_combine_dicts to tsfpga/system_utils.py.
- tsfpga/vivado/build_result.py, tsfpga/vivado/generics.py, and
  tsfpga/vivado/project.py now re-export/import these from their new
  locations, so existing Vivado call sites are unaffected.
- tsfpga/yosys/project.py now only imports from tsfpga.build_result,
  tsfpga.generics, and tsfpga.system_utils, with no reference to
  tsfpga.vivado.

Also fixed while cleaning up:
- Top-level VHDL source file lookup was hardcoded to a "*.vhd" glob,
  breaking for .vhdl/.vho top-level files. Renamed
  _find_module_for_vhdl_entity to _find_vhdl_source_file, which now
  returns the matched HdlFile as well, so the actual resolved path is
  used to look up the compiled source file instead of reconstructing a
  glob with a hardcoded extension.
- Made vhdl_entities a public attribute, consistent with other
  constructor arguments that are documented as being appendable after
  construction (constraints, build_step_hooks, static_generics, ...).
- Quoted paths passed to Yosys's own script commands (read_verilog's
  -I/file arguments, tee -o) to support paths containing spaces. The
  ghdl-yosys-plugin's own argument parsing does not support quoting,
  so the ghdl elaboration command is deliberately left unquoted.
